requirements.txt
streamlit>=1.30.0
PyGithub>=2.1.1
pandas>=2.0.0
requests>=2.31.0
